+++ lib/llamacpp_tools.dart
@@ -1,4 +1,5 @@
 export 'src/docker/docker_builder.dart';
 export 'src/llamacpp_github.dart';
 export 'src/llamacpp_dir.dart';
-export 'src/llamaserver_process.dart';
+export 'src/llamaserver_process.dart' hide PrivateLlamaserverConfigExt;
+export 'src/llamaserver_switcher.dart';