+++ test/llamaserver_process_test.dart
@@ -1,51 +1,23 @@
 import 'dart:convert';
-import 'dart:io';
 import 'package:http/http.dart' as http;
-import 'package:huggingface_tools/huggingface_tools.dart';
-import 'package:llamacpp_tools/src/llamacpp_dir.dart';
 import 'package:llamacpp_tools/src/llamaserver_process.dart';
-import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
-import 'package:llamacpp_tools/src/llamacpp_github.dart';
+import 'test_setup.dart';
 
 void main() {
   group('LllamacppProcess', () {
-    late LlamacppDir llamacppDir;
-    late File modelFile;
+    final testSetup = TestSetup();
 
     setUp(() async {
-      final versions = await LlamacppGithub.scrapeRecentVersions();
-      final targetVersion = versions.first;
-      final llamaPath = path.join(
-        '.dart_tool',
-        'cached',
-        'llama_cpp',
-        targetVersion,
-      );
-      llamacppDir = await LlamacppGithub.downloadAndSetupRelease(
-        targetPath: llamaPath,
-        version: targetVersion,
-      );
-
-      modelFile = File(
-        path.join('.dart_tool', 'cached', 'model', 'gte-small.Q2_K.gguf'),
-      ).absolute;
-      if (!await modelFile.exists()) {
-        await modelFile.parent.create(recursive: true);
-        final model = await getModelInfo('ChristianAzinn/gte-small-gguf');
-
-        final ggufFile = model.ggufModelFiles.firstWhere(
-          (f) => f.filename == 'gte-small.Q2_K.gguf',
-        );
-        await downloadFile(ggufFile, targetFilePath: modelFile.path);
-      }
+      await testSetup.setUp();
     });
 
     test('run', () async {
       final process = LlamaserverProcess(
-        dir: llamacppDir,
-        config: LlamaserverConfig(modelPath: modelFile.path),
+        dir: testSetup.llamacppDir,
+        config: LlamaserverConfig(modelPath: testSetup.modelFile.path),
+        logWriter: (_) {},
       );
       try {
         await process.start();